Day-23/multi-agent-langgraph/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
langgraph>=0.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
